+++ src/main/java/io/github/tato126/practice/common/excetion/post/PostAccessDeniedException.java
@@ -1,4 +1,4 @@
-package io.github.tato126.practice.common.excetion;
+package io.github.tato126.practice.common.excetion.post;
 
 /**
  * 게시글 접근 권한이 없을 때 발생하는 예외입니다.
